Moved protocols to separate files.
